Remove the reference of vsSenOpenGL, use 4level with warning to error for all projects; Make 32bit work, make all input libraries neat and clean Start GLFW_GLEW from scratch with the best OOP design

+++ SLDigitalImageProcess/SLDigitalImageProcess.cpp
@@ -52,7 +52,7 @@ namespace sldip
 		dstImage.CreateLibreImage(srcImage.GetWidth(), srcImage.GetHeight(), srcImage.GetColorType());
 		assert(!dstImage.IsNull());
 
-		memcpy(dstImage.GetBufferEntryForEdit(), srcImage.GetBufferEntry(), srcImage.GetTotalBytes());
+		memcpy(dstImage.GetBufferEntryForEdit(), srcImage.GetBufferEntry(), static_cast<size_t>(srcImage.GetTotalBytes()) );
 	}// End of DuplicateImage(const SLLibreImage& srcImage, SLLibreImage& dstImage)
